+++ modules/device_identity/genrevibes_device_identity_platform/lib/genrevibes_device_identity_platform.dart
@@ -3,3 +3,4 @@ library;
 
 export 'src/att_advertising_id_source.dart';
 export 'src/device_info_vendor_id_source.dart';
+export 'src/platform_advertising_id_source.dart';